+++ aerokit/__init__.py
@@ -1,4 +1,4 @@
 # coding: utf-8
 # import os
 # import sys
-__version__="1.1.0"
+__version__="1.1.2"